+++ src/molgenis/bbmri_eric/publisher.py
@@ -5,7 +5,7 @@ from molgenis.bbmri_eric.enricher import Enricher
 from molgenis.bbmri_eric.errors import EricError, EricWarning
 from molgenis.bbmri_eric.model import NodeData, QualityInfo, Table, TableType
 from molgenis.bbmri_eric.pid_manager import PidManager
-from molgenis.bbmri_eric.pid_service import PidService
+from molgenis.bbmri_eric.pid_service import BasePidService
 from molgenis.bbmri_eric.printer import Printer
 from molgenis.client import MolgenisRequestError
 
@@ -16,7 +16,9 @@ class Publisher:
     public tables.
     """
 
-    def __init__(self, session: EricSession, printer: Printer, pid_service: PidService):
+    def __init__(
+        self, session: EricSession, printer: Printer, pid_service: BasePidService
+    ):
         self.session = session
         self.printer = printer
         self.pid_service = pid_service